refactor(athalar_core): move lib code to core module
Also correct the documentation string about a method in validation reporter.
